Refactor container function runtime into its own package
